fix: add a setup function
- define a setup function to be used as an entry point
- remove `C` and `Config` from the global scope
- import minor updates from upstream
